warn instead of error on ban-types
I added this because when converting middleware to unifiedcicd this
error appeared. The convention is when there are more than a few
instances of an error, we change the error to a warning then add it
here, so other projects in the future can avoid having to override it
in their own local repo settings.